+++ IDisposableAnalyzers/Helpers/Disposable.ShouldDispose.cs
@@ -1,5 +1,6 @@
 ﻿namespace IDisposableAnalyzers
 {
+    using System;
     using System.Threading;
 
     using Gu.Roslyn.AnalyzerExtensions;
@@ -77,6 +78,7 @@
                         => references[0].GetSyntax(cancellationToken),
                     ILocalSymbol { DeclaringSyntaxReferences: { Length: 1 } references }
                         => references[0].GetSyntax(cancellationToken).FirstAncestor<BlockSyntax>(),
+                    _ => throw new InvalidOperationException("Should never get here."),
                 };
             }