+++ ValidCode/Issue273.cs
@@ -1,10 +1,16 @@
-﻿namespace ValidCode
+﻿// ReSharper disable All
+namespace ValidCode
 {
     using System;
 
     public sealed class Issue273 : IDisposable
     {
         private readonly Disposable disposable;
+
+        public Issue273(int x, int y)
+            : this(x + y)
+        {
+        }
 
         public Issue273(int x)
         {